Consolidate agent manifests in pkg/agent
Refactoring. The manifests are not used by the fleet CLI, but were stored in
modules/cli. The manifests were mostly used together.
Move everything into one package, make funcs private.
